requirements.txt
Flask==0.12.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
gunicorn==19.7.1
psycopg2==2.7.3.2
